Unify LMR wrappers into one file (avoiding circular references when importing)
